refactor(mcp): extract risk and response size guard code
Moved RiskClass metadata into risk.go and the result size guard into response_size_guard.go. Kept RiskClass.Has because current production code calls it. Gates: go build ./internal/mcp, go test -race -count=1 ./internal/mcp, go build ./..., go vet ./..., go test -count=1 ./..., golangci-lint run, make catalog-drift.
